fix the return type of multiply_image_and_position
for images of type int8, multiply_image_and_position was limited to positions < 256 (and for int16 to positions < 2^16), because the return type was the same as the image type. However int8 images can have positions larger than that because position indices are of type int32. This commit ensures that the return type is at least int32 so that multiply_image_and_position can handle positions larger than the maximum integer for int8 and int16 images.

closes issue tier1::multiply_image_and_position_func() implicitly assumes that the position has the same type as the input. #287